Porting game.c to SDL2, fixed tab mess.

+++ src/hero.c
@@ -192,7 +192,7 @@ void set_hero_init (struct hero *griel, uint8_t round) {
 
 }
 
-void extralife (struct hero *griel, uint *uplife) {
+void extralife (struct hero *griel, uint8_t *uplife) {
 
   if (griel->score > (5000 + (*uplife * 5000))) {
     *uplife += 1;
@@ -202,7 +202,7 @@ void extralife (struct hero *griel, uint *uplife) {
 
 }
 
-void show_hero (struct hero *griel, int counter, SDL_Surface *window, SDL_Surface *blocks, int *round, int *step, uint *waittime, uint *soundblock, Mix_Chunk *giveup) {
+void show_hero (struct hero *griel, uint8_t counter, SDL_Surface *window, SDL_Surface *blocks, int *round, int *step, uint16_t *waittime, uint8_t *soundblock, Mix_Chunk *giveup) {
 
   SDL_Rect srchero = {96,32,16,16};
   SDL_Rect desthero = {0,0,16,16};
